+++ apps/earn-protocol/hooks/useVisibleParagraph.ts
